Convert LWRP to a custom resource
